feat: add tests for icons package [MDS-55] (#1683)
